+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2022-05-16
+
+### Added
+
+- Build argument `PHP_VERSION` for Docker container (defaults to `7.4`)
+
+### Changed
+
+- Switch to `main` branch for tests of `composer-update-reporter`
+
+### Fixed
+
+- Respect platform requirements during update check
+- Use test application variants for different PHP major versions
+
 ## [1.2.0] - 2022-04-13
 
 ### Added
@@ -267,7 +282,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/eliashaeussler/composer-update-check/compare/1.2.0...develop
+[Unreleased]: https://github.com/eliashaeussler/composer-update-check/compare/1.3.0...develop
+[1.3.0]: https://github.com/eliashaeussler/composer-update-check/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/eliashaeussler/composer-update-check/compare/1.1.3...1.2.0
 [1.1.3]: https://github.com/eliashaeussler/composer-update-check/compare/1.1.2...1.1.3
 [1.1.2]: https://github.com/eliashaeussler/composer-update-check/compare/1.1.1...1.1.2